Create SmallHigh.py
Given a sorted array of integers, return the smallest and highest index of the given key. 
You must return -1 if the indexes are not found. 
The array length can be in the millions with many duplicates.
e,g, if input= 1 3 6 6 
for key 1: smallest=0,highest=0
for key 3: smallest=1,highest=1
for key 6 smallest=3,highest=2